helpers/llm/requirements.txt
boto3==1.42.34
python-dotenv==1.1.0